Add files via upload
This is the original biryani house code. This has the correct billing system but a very complicated delete system. (here user has to select how many different items they want to delete and then for loop will be executed that many times and user will have to give input for the respective number to delete) huge ordeal not convinient!